src/requirements.txt
Django~=2.0.6
django-admin-sortable2
django-flat-responsive
django-image-cropping
django-import-export
django-modeltranslation
django-phonenumber-field
django-unused-media
easy-thumbnails
django-ckeditor
gunicorn==19.7.1
